Creation de classes de calibrage pour la video probleme de memoire regle le traitement actuel consiste en deux filtres medians

+++ src/traitementVideo/VirtualPixel.java
@@ -37,7 +37,7 @@ public class VirtualPixel {
 		this.brightness = brightness;
 	}
 
-	public int getIntensite() {
+	public byte getIntensite() {
 		return intensite;
 	}
